keeping hook in Game
