+++ server/db/models/Game.js
@@ -1,8 +1,6 @@
 const Sequelize = require('sequelize');
 const db = require('../db');
 const Path = require('./path');
-const Scores = require('./scores');
-const User = require('./User');
 
 const Game = db.define('game', {
   stage: {
@@ -32,36 +30,5 @@ Game.addHook('beforeValidate', async (game, options) => {
     await game.save();
   }
 });
-
-// Game.createGame = async function (userId, path_name) {
-//   let game = await Game.create({
-//     path_name,
-//     userId,
-//   });
-//   await Scores.create({
-//     gameId: game.id,
-//   });
-
-//   game = await Game.findOne({
-//     where: {
-//       userId: user.id,
-//       status: 'ingame',
-//     },
-//     include: Scores,
-//   });
-//   return game;
-// };
-
-Game.getLeadership = function () {
-  const leadership = Game.findAll({
-    where: {
-      status: 'finished',
-    },
-    include: [Scores, User],
-    order: [[{ model: Scores }, 'total_score', 'DESC']],
-    limit: 10,
-  });
-  return leadership;
-};
 
 module.exports = Game;
